+++ examples/data-sources/meraki_camera_custom_analytics/data-source.tf
@@ -1,0 +1,3 @@
+data "meraki_camera_custom_analytics" "example" {
+  serial = "1234-ABCD-1234"
+}