+++ examples/resources/meraki_camera_custom_analytics/import.sh
@@ -1,0 +1,1 @@
+terraform import meraki_camera_custom_analytics.example "<serial>"