+++ gen/definition.go
@@ -66,7 +66,7 @@ data "meraki_organization" "test" {
 resource "meraki_network" "test" {
   organization_id = data.meraki_organization.test.id
   name            = var.test_network
-  product_types   = ["switch", "wireless", "appliance", "sensor"]
+  product_types   = ["switch", "wireless", "appliance", "sensor", "camera"]
 }
 `
 
@@ -77,7 +77,7 @@ data "meraki_organization" "test" {
 resource "meraki_network" "test" {
   organization_id = data.meraki_organization.test.id
   name            = var.test_network
-  product_types   = ["switch", "wireless", "appliance", "sensor"]
+  product_types   = ["switch", "wireless", "appliance", "sensor", "camera"]
 }
 resource "meraki_network_device_claim" "test" {
   network_id = meraki_network.test.id