+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ## 0.1.7 (unreleased)
 
 - Add `meraki_organization_auth_radius_server` resource and data sources
+- Add `meraki_camera_custom_analytics` resource and data source
 
 ## 0.1.6